server, client: Distribute latency to clients
This allows us to distribute the overall pipeline latency from the
server to all clients. This is important for two cases -- the first is
when we're on heterogenous devices that might have different total
latencies (e.g. default audio sink latency is different), and the second
is that network latencies might vary for live pipelines, and we want the
ability to distribute a common total latency to all clients if needed.

The latency is configurable via a property on the server, and the
example allows setting this as well now.