client: Change API so client doesn't have to create playbin
It's probably more natural, API-wise, for us to create the element we
expect to have and the client can access it and configure it as
required.

We should move this to use playbin3 before long as that is what will be
recommended in the future. Not doing this now to allow testing on
devices with GStreamer < 1.10.