server, client: Implement basic playlist support
This adds support for playlists to the server and client. The clients
need some improvement to use the playlists optimally (e.g. skip over
tracks that have already finished playing rather than loading and then
skipping them).

In the process, we change GstSyncServerInfo to only work with playlists,
with the URI being a special-case playlist (this should probably just go
away entirely).

Track durations in the playlist are not currently used, but can be for
the use-case mentioned above.

Another feature to be added is a delay between streams -- this can help
provide smoother transitions between tracks since different devices
might load media with different delays, depending on the URI and storage
media (http vs. files vs. files on slow media).