server, client: Implement API to set per-client transformations
The transformations are sent via a dictionary mapping client IDs to the
desired transformations, and are applied as:

  videocrop ! videoflip ! videoscale ! videobox

This obviously doesn't work for GL-based systems, or ones where some or
all of these operations can be performed by the sink. We need to figure
out a way to do this correctly so it works in either case.